Разделит main.js на разные модули
